Rich, I added a method to retrieve lesson recommendations for a user. You can use it on the user's home page to display a carousel of recommended lessons.
Use it with the following

User.lesson_recommendations(<user>, <size>)

where user is some user and size is the number of recommendations you want (default is 5).

It first looks to see what other users have purchased who have also purchased same videos the user purchased.
If that is less then the size specified, then it will then retrieve the highest rated videos as required to fill out the
number specified